requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
scikit-learn==0.24.2
joblib==1.0.2
markdown==3.3.4 